refactor: rearrange server routes
