+++ apps/server/src/cors/auth0.ts
