+++ apps/server/src/cors/index.ts
@@ -1,8 +1,7 @@
 import type { CorsOptions } from 'cors'
-import { authZeroIpAddresses } from './auth0'
 
 const origins = ['https://modtree.vercel.app', 'http://localhost:3000']
 
 export const corsOpts: CorsOptions = {
-  origin: [...authZeroIpAddresses, ...origins],
+  origin: [...origins],
 }
