+++ package.json
@@ -34,6 +34,7 @@
     "chalk": "^4.1.2",
     "core-js": "^3.6.5",
     "cors": "^2.8.5",
+    "d3": "^7.4.4",
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
     "express-validator": "^6.14.1",
@@ -72,6 +73,7 @@
     "@tailwindcss/typography": "^0.5.2",
     "@testing-library/react": "12.1.5",
     "@types/cors": "^2.8.12",
+    "@types/d3": "^7.4.0",
     "@types/express": "^4.17.13",
     "@types/inquirer": "^8.2.1",
     "@types/jest": "27.4.1",
@@ -113,5 +115,8 @@
     "npm": "8.x",
     "node": "16.x",
     "yarn": "1.x"
+  },
+  "resolutions": {
+    "@types/react": "17.0.37"
   }
 }
