+++ tsconfig.base.json
@@ -13,8 +13,8 @@
     "target": "es6",
     "module": "commonjs",
     "lib": ["es5", "es6", "dom"],
-    "skipLibCheck": true,
-    "skipDefaultLibCheck": true,
+    "skipLibCheck": false,
+    "skipDefaultLibCheck": false,
     "baseUrl": ".",
     "paths": {
       "@modtree/entity": ["libs/entity/src/index.ts"],
